# src/app.rs
use crate::config::Config;
use crate::{config, handler};
use axum::Router;
use tokio::net::TcpListener;
use tokio::signal;
use utoipa::OpenApi;

#[derive(OpenApi)]
#[openapi(info(
    title = "Token Exchange as a Service (Texas)",
    description = "Texas implements OAuth token fetch, exchange, and validation, so that you don't have to.",
    contact(name = "Nais", url = "https://nais.io")
))]
pub struct App {
    router: Router,
    pub listener: TcpListener,
    pub probe_listener: Option<TcpListener>,
}

#[derive(thiserror::Error, Debug)]
pub enum Error {
    #[error("set up listening socket: {0}")]
    BindAddress(std::io::Error),
    #[error("describe socket local address: {0}")]
    LocalAddress(std::io::Error),
    #[error("{0}")]
    InitHandler(handler::InitError),
    #[error("invalid configuration: {0}")]
    Configuration(config::Error),
}

impl App {
    pub async fn new_from_env() -> Result<Self, Error> {
        let cfg = Config::new_from_env().map_err(Error::Configuration)?;
        Self::new_from_config(cfg).await
    }

    pub async fn new_from_config(cfg: Config) -> Result<Self, Error> {
        let bind_address = cfg.bind_address.clone();
        let listener = TcpListener::bind(bind_address).await.map_err(Error::BindAddress)?;
        let api_address = listener.local_addr().map_err(Error::LocalAddress)?;
        log::info!("Serving API on http://{api_address:?}");
        #[cfg(feature = "openapi")]
        log::info!(
            "Swagger API documentation: http://{:?}/swagger-ui",
            api_address
        );

        let probe_listener = if let Some(addr) = cfg.probe_bind_address.as_ref() {
            let listener = TcpListener::bind(addr).await.map_err(Error::BindAddress)?;
            let probe_address = listener.local_addr().map_err(Error::LocalAddress)?;
            log::debug!("Serving probes on http://{probe_address:?}");
            Some(listener)
        } else {
            None
        };

        let state = handler::State::from_config(cfg).await.map_err(Error::InitHandler)?;

        #[cfg(not(feature = "openapi"))]
        let router = || -> Router {
            let (router, _) = router::api(state);
            router
        };
        #[cfg(feature = "openapi")]
        let router = || -> Router {
            use utoipa_swagger_ui::SwaggerUi;

            let (router, openapi) = router::api(state);
            let swagger =
                SwaggerUi::new("/swagger-ui").url("/api-docs/openapi.json", openapi.clone());
            router.merge(swagger)
        };

        Ok(Self {
            router: router(),
            listener,
            probe_listener,
        })
    }

    pub async fn run(self) {
        if let Some(probe_listener) = self.probe_listener {
            let api_handler = tokio::task::spawn(async move {
                serve(self.listener, self.router).await;
            });
            let probe_handler = tokio::task::spawn(async move {
                serve(probe_listener, router::probe()).await;
            });
            let _ = tokio::try_join!(api_handler, probe_handler);
        } else {
            serve(self.listener, self.router).await;
        }

        log::debug!("Texas shut down gracefully");
    }
}

async fn serve(listener: TcpListener, router: Router) {
    // axum::serve doesn't actually return an error according to the documentation, so we unwrap here
    axum::serve(listener, router).with_graceful_shutdown(shutdown_signal()).await.unwrap();
}

async fn shutdown_signal() {
    let ctrl_c = async {
        signal::ctrl_c().await.expect("failed to install Ctrl+C handler");
    };

    #[cfg(unix)]
    let terminate = async {
        signal::unix::signal(signal::unix::SignalKind::terminate())
            .expect("failed to install SIGTERM handler")
            .recv()
            .await;
    };
    #[cfg(not(unix))]
    let terminate = std::future::pending::<()>();

    tokio::select! {
        () = ctrl_c => log::debug!{"Received Ctrl+C / SIGINT"},
        () = terminate => log::debug!{"Received SIGTERM"},
    }
}

pub mod router {
    use crate::app::App;
    use crate::handler;
    use crate::tracing::record_http_response_secs;
    use axum::Router;
    use axum::extract::MatchedPath;
    use axum::http::{Request, StatusCode};
    use axum::response::Response;
    use axum::routing::get;
    use opentelemetry::baggage::BaggageExt;
    use opentelemetry::{KeyValue, global};
    use opentelemetry_http::HeaderExtractor;
    use std::time::Duration;
    use tower_http::trace::TraceLayer;
    use tracing::{Span, field, info_span};
    use tracing_opentelemetry::OpenTelemetrySpanExt;
    use utoipa::{OpenApi, openapi};
    use utoipa_axum::router::OpenApiRouter;
    use utoipa_axum::routes;

    pub fn api(state: handler::State) -> (Router, openapi::OpenApi) {
        let trace_layer = TraceLayer::new_for_http()
            .make_span_with(move |request: &Request<_>| {
                // Log the matched route's path (with placeholders not filled in).
                // Use request.uri() or OriginalUri if you want the real path.
                let path = request.extensions().get::<MatchedPath>().map(MatchedPath::as_str);

                let root_span = info_span!(
                    "Handle incoming request",
                    "http.request.method" = ?request.method(),
                    "http.response.status_code" = field::Empty, // to be populated in on_response
                    "http.route" = path,
                    "http.version" = ?request.version(),
                    "otel.kind" = "server",
                );

                let parent_context = global::get_text_map_propagator(|propagator| {
                    propagator.extract(&HeaderExtractor(request.headers()))
                });
                let context = parent_context.with_baggage(vec![KeyValue::new(
                    "path".to_string(),
                    path.unwrap_or_default().to_string(),
                )]);
                let _ = root_span.set_parent(context.clone());
                root_span
            })
            .on_response(move |response: &Response, latency: Duration, span: &Span| {
                let path = span
                    .context()
                    .baggage()
                    .get("path")
                    .map(ToString::to_string)
                    .unwrap_or_default();
                span.record("http.response.status_code", response.status().as_u16());
                record_http_response_secs(&path, latency, response.status());
            });

        let api = OpenApiRouter::default()
            .routes(routes!(handler::token))
            .routes(routes!(handler::token_exchange))
            .routes(routes!(handler::token_introspect))
            .layer(trace_layer)
            .with_state(state);

        OpenApiRouter::with_openapi(App::openapi()).merge(api).split_for_parts()
    }

    pub(super) fn probe() -> Router {
        async fn healthz() -> (StatusCode, &'static str) {
            (StatusCode::OK, "ok")
        }

        Router::new().route("/", get(healthz)).route("/healthz", get(healthz))
    }
}